drm: bridge: synopsis: add dsi v1.01 support
The Synopsis MIPI DSI v1.01 host controller is quite widely used
on platforms like i.mx6 and is not very different from the other
versions like the 1.31/1.30 used on rockchip/stm. The protocols
appear to be the same, only the register layout is different and
the newer versions have new features symbolized by new registers
so adding support for it is just a matter of defining the new
layout and adding a couple of dsi version checks.

Change-Id: Icc6b455bff830d570e3b1713198032cc8a9a9776
Signed-off-by: Adrian Ratiu <adrian.ratiu@collabora.com>
---
Changes since v5:
  - Fixed cfg_phy_status range from [0,0] to [0,2]

New in v5.