drm: bridge: dw-mipi-dsi: split low power cfg register into fields
According to the Host Registers documentation for IMX, STM and RK
the LP cfg register should not be written entirely in one go because
some bits are reserved and should be kept to reset values, for eg.
BIT(15) which is reserved in all versions.

This also cleans up the code by removing the the ugly defines
and making field ranges & values written more explicit.

Change-Id: I294339d42d6fd8030da1e70923c9da01b60d70a4
Signed-off-by: Adrian Ratiu <adrian.ratiu@collabora.com>
---
New in v6.